Model rework
Separated attendee, event, housing, and restaurant models into their own files

Changed schema imports in apis to match with the changes
